# openapi.json
{
  "components": {
    "schemas": {
...
      "CallSummarySettingsIO": {
        "properties": {
          "emails": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Emails",
            "type": "array"
          },
          "enabled": {
            "default": false,
            "title": "Enabled",
            "type": "boolean"
          },
          "include_transcript": {
            "default": false,
            "title": "Include Transcript",
            "type": "boolean"
          },
          "min_seconds": {
            "default": 15,
            "maximum": 600.0,
            "minimum": 0.0,
            "title": "Min Seconds",
            "type": "integer"
          },
          "timezone": {
            "default": "Asia/Kolkata",
            "maxLength": 64,
            "title": "Timezone",
            "type": "string"
          },
          "to_assignee": {
            "default": true,
            "title": "To Assignee",
            "type": "boolean"
          },
          "to_managers": {
            "default": true,
            "title": "To Managers",
            "type": "boolean"
          }
        },
        "title": "CallSummarySettingsIO",
        "type": "object"
      },
      "CallSummarySettingsOut": {
        "properties": {
          "email_ready": {
            "title": "Email Ready",
            "type": "boolean"
          },
          "emails": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Emails",
            "type": "array"
          },
          "enabled": {
            "default": false,
            "title": "Enabled",
            "type": "boolean"
          },
          "include_transcript": {
            "default": false,
            "title": "Include Transcript",
            "type": "boolean"
          },
          "min_seconds": {
            "default": 15,
            "maximum": 600.0,
            "minimum": 0.0,
            "title": "Min Seconds",
            "type": "integer"
          },
          "timezone": {
            "default": "Asia/Kolkata",
            "maxLength": 64,
            "title": "Timezone",
            "type": "string"
          },
          "to_assignee": {
            "default": true,
            "title": "To Assignee",
            "type": "boolean"
          },
          "to_managers": {
            "default": true,
            "title": "To Managers",
            "type": "boolean"
          }
        },
        "required": [
          "email_ready"
        ],
        "title": "CallSummarySettingsOut",
        "type": "object"
      },
      "CallSummaryTestOut": {
        "properties": {
          "recipients": {
            "items": {
              "type": "string"
            },
            "title": "Recipients",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "recipients",
          "status"
        ],
        "title": "CallSummaryTestOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkspaceOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "plan": {
            "title": "Plan",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "plan",
          "status"
        ],
        "title": "WorkspaceOut",
        "type": "object"
      },
      "WorkspaceUpdate": {
        "properties": {
          "name": {
            "maxLength": 150,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "WorkspaceUpdate",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "Production-ready FastAPI backend supporting user authorization, lead records, and calls metrics with a PostgreSQL database.",
    "title": "aVn Agent API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/workspace": {
      "get": {
        "operationId": "get_workspace_api_workspace_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Workspace",
        "tags": [
          "Workspace"
        ]
      },
      "patch": {
        "operationId": "update_workspace_api_workspace_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Workspace",
        "tags": [
          "Workspace"
        ]
      }
    },
    "/api/workspace/call-summaries": {
      "get": {
        "description": "Who gets an email after each call.",
        "operationId": "get_call_summaries_api_workspace_call_summaries_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallSummarySettingsOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Call Summaries",
        "tags": [
          "Workspace"
        ]
      },
      "put": {
        "operationId": "update_call_summaries_api_workspace_call_summaries_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CallSummarySettingsIO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallSummarySettingsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Call Summaries",
        "tags": [
          "Workspace"
        ]
      }
    },
    "/api/workspace/call-summaries/test": {
      "post": {
        "description": "Email the latest finished call's summary to the configured recipients now (the call keeps its own status).",
        "operationId": "test_call_summary_api_workspace_call_summaries_test_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallSummaryTestOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Test Call Summary",
        "tags": [
          "Workspace"
        ]
      }
    },
...
  }
}